streamlit/requirements.txt
altair==5.2.0
asn1crypto==1.5.1
async-timeout==4.0.3
asyncpg==0.29.0
atpublic==4.0
attrs==23.1.0
blinker==1.7.0
blis==0.7.11
boto3==1.33.11
botocore==1.33.11
branca==0.7.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==2.2.1
confection==0.1.4
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
cymem==2.0.8
dnspython==2.4.2
ecdsa==0.18.0
filelock==3.13.1
folium==0.15.1
fonttools==4.46.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==3.11.0
plotly==5.18.0
preshed==3.0.9
protobuf==4.25.1
pyarrow==14.0.1
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
s3transfer==0.8.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
snowflake==0.5.1
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.11.1
snowflake._legacy==0.4.0
snowflake.core==0.5.1
sortedcontainers==2.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.29.0
streamlit-folium==0.17.4
tenacity==8.2.3
thinc==8.2.1
threadpoolctl==3.2.0
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
weasel==0.3.4
xyzservices==2023.10.1
zipp==3.17.0
zstandard==0.22.0
